Move utility methods
