+++ xdsl/transforms/loop_hoist_memref.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from xdsl.dialects import builtin, memref, scf
-from xdsl.ir import Block, Operation, Region
+from xdsl.ir import Block, Operation, Region, SSAValue
 from xdsl.irdl import Operand
 from xdsl.passes import MLContext, ModulePass
 from xdsl.pattern_rewriter import (
@@ -11,11 +11,62 @@ from xdsl.pattern_rewriter import (
     RewritePattern,
     op_type_rewrite_pattern,
 )
-from xdsl.transforms.utils import (
-    find_same_target_store,
-    get_operation_at_index,
-    is_loop_dependent,
-)
+
+
+# TODO replace by functionality (when added) as described in https://github.com/xdslproject/xdsl/issues/2128
+def get_operation_at_index(block: Block, idx: int) -> Operation:
+    """Get an operation by its position in its parent block."""
+
+    for _idx, block_op in enumerate(block.ops):
+        if idx == _idx:
+            return block_op
+
+    raise ValueError(
+        f"Cannot get operation by out-of-bounds index {idx} in its parent block."
+    )
+
+
+def find_same_target_store(load: memref.Load):
+    parent_block = load.parent_block()
+
+    if parent_block is None:
+        return None
+
+    found_op = None
+
+    for op in parent_block.ops:
+        if (
+            isinstance(op, memref.Store)
+            and op.memref == load.memref
+            and op.indices == load.indices
+        ):
+            if found_op is None:
+                found_op = op
+            else:
+                return None
+
+    return found_op
+
+
+def is_loop_dependent(val: SSAValue, loop: scf.For):
+    worklist: set[SSAValue] = set()
+    visited: set[SSAValue] = set()
+
+    worklist.add(val)
+
+    while worklist:
+        val = worklist.pop()
+        visited.add(val)
+
+        if val is loop.body.block.args[0]:
+            return True
+
+        if isinstance(val.owner, Operation):
+            for oprnd in val.owner.operands:
+                if oprnd not in visited:
+                    worklist.add(oprnd)
+
+    return False
 
 
 @dataclass
